Moved ADB stream tests into own class
